+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.4.2
+### Added
+- Added lookup api test in aas-registry-e2e-test action
+## fixed
+- KICS findings fixed
+- Fixed behaviour of access rule management in case of multiple specificAssetIds with same keys and different values.
+- Update Springboot to version 3.2.4
+
 ## 0.4.1
 ### Added
 - Added API POST lookup/shellsByAssetLink to retrieve shell ids without base64 encryption.